+++ src/frontend/apps/web/src/features/auth/index.ts
@@ -1,2 +1,3 @@
 export { LoginButton, LogoutButton } from './ui';
 export { requestLogin } from './api/requestLogin.api';
+export { useLoginRedirect } from './model/use-login-redirect';
